rework read()/write() to be numpy-friendly
I was mistaken about GxFpga{Read,Write}{Register,Memory} being able
to work with non-dword quantities, they actually throw an error if
either the address or the size is not a multiple of 4. So change
the size argument to read() to be a count for a read-multiple
operation that creates a numpy.ndarray and make write() accept
anything coercable to a numpy.ndarray.